Move code from LightupCompilationWithAnalyzers to CompilationWithAnalyzersExtensions; fix comments.
